Switch from RSL to RSL_LITE
Tested with ibm, gnu and pgi compilers on bluevista, khair, hazel and big on con200, afwa_swa and
kma_t213 test cases.

All identical apart from the thinning option for radiances, which appears to be sensitive to the
exact positioning of domain boundaries. As the problem is with the thinning, not the RSL_LITE,
the afwa_swa regression test has thinning switched off until the problem is resolved

Thanks to Mike for solving some nasty bugs in the RSL_LITE code


git-svn-id: https://svn-wrf-var.cgd.ucar.edu/trunk@3056 70815b0f-b814-0410-81f7-dfa10a885da6