Reintroduction of wrf compilation system, part 2
To allow a fair fight between compilation systems, I've reintroduced the 
wrf compilation system in the same place as it is in the wrf repository,
and moved the wrfvar system into the build directory. They both compile
everything, wrf and wrfvar

1) If you want to continue with the wrfvar system

cd build
./clean -a
./configure
./compile all_wrfvar

2) If you want to use the wrf system (though God knows why)

./clean -a
./configure wrfvar (Only choose options that say WRFVAR)
./compile all_wrfvar

Note that like wrf trunk, there are problems with the g95 compiler, and
our old faithful pgi compiler in RSL_LITE mode is not a wrf standard, so
is also unavailable for wrfvar

Tested fully on bluevista, other systems may need more work.


git-svn-id: https://svn-wrf-var.cgd.ucar.edu/trunk@3169 70815b0f-b814-0410-81f7-dfa10a885da6